pyproject.toml
[project]
name = "wantedocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.0.1",
    "langchain==0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "langgraph>=1.0.1",
    "openai>=2.5.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
